ref+fix: brush is no longer highlightning while image is moved, Config - create interface, positions - calculate positionsfor interface, windows - add file/save file
