+++ App/CMakeLists.txt
@@ -9,28 +9,25 @@ set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_SOURCE_DIR})
 
 set(SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/main.cpp
+
     ${CMAKE_CURRENT_SOURCE_DIR}/Back/Back.cpp
+
     ${CMAKE_CURRENT_SOURCE_DIR}/Brush/Brush.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/Button/Button.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/FileField/FileField.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/Image/Image.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/MainWindow/Interface/Interface.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/StatusBar/StatusBar.cpp
-)
+     ${CMAKE_CURRENT_SOURCE_DIR}/StatusBar/StatusBar.cpp
 
-set(HEADERS
-    ${CMAKE_CURRENT_SOURCE_DIR}/Back/Back.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Brush/Brush.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Button/Button.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/FileField/FileField.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Image/Image.hpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/MainWindow/MainWindow.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/MainWindow/Interface/Interface.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/StatusBar/StatusBar.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/MainWindow/Interface/Config/Config.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/MainWindow/Interface/Windows/Windows.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/MainWindow/Interface/Positions/Positions.cpp
+
+
 )
 
 find_package(SFML 2.5 COMPONENTS system window graphics REQUIRED)
 
-add_executable(ImageEditor ${SOURCES} ${HEADERS})
+add_executable(ImageEditor ${SOURCES})
 
 target_link_libraries(ImageEditor sfml-system sfml-window sfml-graphics)
